refactor(experimental): move rpc-transport type tests into `src`
